refactor main/parse into scripts.pixsim
